# backend-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: Matchlock Backend API
  version: 0.1.0
  description: |
    REST API served by `backend-go` (keeper process). Match data is ingested from TxLINE SSE
    and cached in Redis. Wager accounts are read from Solana RPC.

    **CORS:** `GET` and `OPTIONS` are allowed from origins in `CORS_ALLOWED_ORIGINS`
    (default `http://localhost:5173`, `http://127.0.0.1:5173`).

    **On-chain writes** (`make_wager`, `accept_wager`, `cancel_wager`) are signed in the
    browser wallet via the Solana program — this API is read-only for on-chain state.

    **Devnet defaults** (see `backend-go/.env`):
    - Program: `B39Vk22T2VPpqBEbGkW51BzFC6sNeeiQQ1mqwdvCJ2H4`
    - Stablecoin mint: `ELWTKspHKCnCfCiCiqYw1EDH77k8VCP74dK9qytG2Ujh`
    - Cluster: `devnet`
servers:
  - url: http://localhost:8080
    description: Local keeper (`HTTP_ADDR`, default `:8080`)
  - url: http://127.0.0.1:8080
    description: Local keeper (loopback)
...
paths:
...
  /matches:
    get:
      tags: [matches]
      operationId: listMatches
      summary: List cached matches
      description: |
        Returns all matches currently in the Redis cache, populated by the TxLINE scores
        SSE ingestor. Empty array when no fixtures have been observed yet.
      responses:
        "200":
          description: Match list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchView"
        "500":
          $ref: "#/components/responses/CacheError"
  /matches/stream:
    get:
      tags: [matches]
      operationId: matchStream
      summary: SSE match update stream
      description: |
        Server-Sent Events stream of real-time match score updates.
        Each event is a JSON-encoded `MatchView` payload.

        A heartbeat comment (`:`) is sent every 15 seconds to keep the connection alive.
      responses:
        "200":
          description: SSE event stream of match updates
          content:
            text/event-stream:
              schema:
                type: string
                description: "`data: <MatchView>\\n\\n`"
        "500":
          description: SSE unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: match streaming not configured
                code: SSE_UNAVAILABLE
  /matches/{id}:
    get:
      tags: [matches]
      operationId: getMatch
      summary: Get match by ID
      parameters:
        - $ref: "#/components/parameters/MatchId"
      responses:
        "200":
          description: Match found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchView"
        "400":
          $ref: "#/components/responses/InvalidMatchId"
        "404":
          $ref: "#/components/responses/MatchNotFound"
        "500":
          $ref: "#/components/responses/CacheError"
  /wagers:
    get:
      tags: [wagers]
      operationId: listWagers
      summary: List on-chain wagers
      description: |
        Scans Matchlock program accounts via Solana RPC with optional filters.

        When no `status` or `wallet` filter is supplied, only `open` and `matched` wagers
        are returned (market browsing). Use `wallet` to query a user's full history.
      parameters:
        - $ref: "#/components/parameters/MatchIdQuery"
        - $ref: "#/components/parameters/StatusQuery"
        - $ref: "#/components/parameters/WalletQuery"
      responses:
        "200":
          description: Wager list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WagerView"
        "400":
          $ref: "#/components/responses/InvalidQuery"
        "502":
          $ref: "#/components/responses/RpcError"
  /wagers/history:
    get:
      tags: [wagers]
      operationId: listWagerHistory
      summary: List wallet bet history
      description: |
        Returns wallet-specific wager history with backend-owned settlement, outcome,
        and date filtering. Match metadata is embedded for the history UI.
      parameters:
        - $ref: "#/components/parameters/HistoryWalletQuery"
        - $ref: "#/components/parameters/HistorySettlementStatusQuery"
        - $ref: "#/components/parameters/HistoryOutcomeQuery"
        - $ref: "#/components/parameters/HistoryFromQuery"
        - $ref: "#/components/parameters/HistoryToQuery"
        - $ref: "#/components/parameters/OffsetQuery"
        - $ref: "#/components/parameters/LimitQuery"
      responses:
        "200":
          description: Wager history page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WagerHistoryPageView"
        "400":
          $ref: "#/components/responses/InvalidQuery"
        "500":
          $ref: "#/components/responses/CacheError"
        "502":
          $ref: "#/components/responses/RpcError"
  /wagers/{pubkey}:
    get:
      tags: [wagers]
      operationId: getWager
      summary: Get wager by account pubkey
      parameters:
        - $ref: "#/components/parameters/WagerPubkey"
      responses:
        "200":
          description: Wager found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WagerView"
        "400":
          $ref: "#/components/responses/InvalidPubkey"
        "404":
          $ref: "#/components/responses/WagerNotFound"
        "502":
          $ref: "#/components/responses/RpcError"
  /wagers/{pubkey}/settlement:
    get:
      tags: [settlement]
      operationId: getWagerSettlement
      summary: Get settlement status for a matched wager
      description: |
        Returns the current settlement state of a matched wager. States include:
        `match_live`, `match_ended_unverified`, `claimable`, `refundable`, `queued`,
        `retrying`, `settled`, `failed`, and `not_applicable`.
      parameters:
        - $ref: "#/components/parameters/WagerPubkey"
      responses:
        "200":
          description: Settlement status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementStatusView"
        "400":
          $ref: "#/components/responses/InvalidPubkey"
        "502":
          $ref: "#/components/responses/RpcError"
  /wagers/{pubkey}/settlement-proof:
    get:
      tags: [settlement]
      operationId: getWagerSettlementProof
      summary: Build settlement proof for a matched wager
      description: |
        Builds the on-chain settlement proof data (Merkle proofs, stat validation) needed
        by the winner's wallet to call `settle_wager` on the Matchlock program.
      parameters:
        - $ref: "#/components/parameters/WagerPubkey"
      responses:
        "200":
          description: Settlement proof payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SettlementProof"
        "400":
          $ref: "#/components/responses/InvalidPubkey"
        "404":
          $ref: "#/components/responses/WagerNotFound"
        "409":
          description: Match not final or wager ineligible
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                matchNotFinal:
                  summary: Match is still live
                  value:
                    error: match is not final yet
                    code: MATCH_NOT_FINAL
                invalidStatus:
                  summary: Wager not in matched state
                  value:
                    error: wager is not eligible for settlement
                    code: INVALID_STATUS
        "502":
          $ref: "#/components/responses/RpcError"
        "503":
          description: Settlement proof service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: settlement proof service unavailable
                code: NOT_READY
  /auth/magic-link:
    post:
      tags: [auth]
      operationId: postMagicLink
      summary: Request email magic-link
      description: |
        Sends a sign-in link to the given email address. If the email is not registered,
        an account is created automatically (silent sign-up).

        Rate-limited to 8 requests per hour per email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MagicLinkRequest"
      responses:
        "204":
          description: Magic link sent (no content)
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid email address
                code: INVALID_EMAIL
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: please wait a minute before requesting another sign-in link (max 8 per hour)
                code: RATE_LIMITED
  /auth/verify:
    get:
      tags: [auth]
      operationId: getVerifyMagicLink
      summary: Verify magic-link token
      description: |
        Exchanges a magic-link token for an authenticated session. Sets `access_token`
        and `refresh_token` HTTP-only cookies and returns the user profile.
      parameters:
        - $ref: "#/components/parameters/TokenQuery"
      responses:
        "200":
          description: Session created
          headers:
            Set-Cookie:
              description: access_token and refresh_token cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Link expired or already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: link expired or already used
                code: INVALID_TOKEN
        "500":
          description: Auth error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: could not verify login
                code: AUTH_ERROR
  /auth/refresh:
    post:
      tags: [auth]
      operationId: postRefresh
      summary: Refresh session token
      description: |
        Uses the `refresh_token` cookie to issue new access and refresh tokens.
        Returns the user profile and sets new cookies.
      responses:
        "200":
          description: Session refreshed
          headers:
            Set-Cookie:
              description: New access_token and refresh_token cookies
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Expired or missing refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: session expired; sign in again
                code: UNAUTHORIZED
...
  /auth/me:
    get:
      tags: [auth]
      operationId: getMe
      summary: Get current user profile
      description: Returns the authenticated user's profile and linked wallets.
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/AuthError"
    patch:
      tags: [auth]
      operationId: patchMe
      summary: Update display name
      description: Updates the authenticated user's public display name (3-32 chars, alphanumeric + underscore).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateDisplayNameRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Invalid display name
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: username must be 3-32 characters (letters, numbers, underscore)
                code: INVALID_DISPLAY_NAME
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/AuthError"
  /auth/wallets/check:
    get:
      tags: [auth]
      operationId: getWalletCheck
      summary: Check wallet binding status
      description: |
        Checks whether a Solana wallet pubkey is unlinked, linked to the current user,
        or owned by another account.
      parameters:
        - $ref: "#/components/parameters/PubkeyQuery"
      responses:
        "200":
          description: Wallet binding status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBindingView"
        "400":
          description: Invalid pubkey
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: pubkey required
                code: INVALID_PUBKEY
        "401":
          $ref: "#/components/responses/Unauthorized"
  /auth/wallets/challenge:
    post:
      tags: [auth]
      operationId: postWalletLinkChallenge
      summary: Create wallet link challenge
      description: |
        Generates a sign-in message that the user's wallet must sign to prove ownership.
        The returned message should be presented to the user's Solana wallet for signing.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletLinkChallengeRequest"
      responses:
        "200":
          description: Challenge message created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletLinkChallengeResponse"
        "400":
          description: Challenge creation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid pubkey
                code: CHALLENGE_FAILED
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Wallet already linked to another account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: this wallet is already linked to another Matchlock account
                code: WALLET_OWNED_BY_OTHER
  /auth/wallets/link:
    post:
      tags: [auth]
      operationId: postWalletLink
      summary: Link a wallet to the user account
      description: |
        Completes wallet linking by providing the signed challenge message.
        The signature proves ownership of the Solana wallet.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WalletLinkRequest"
      responses:
        "200":
          description: Wallet linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBindingView"
        "400":
          description: Wallet link failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid signature
                code: WALLET_LINK_FAILED
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Wallet already linked to another account
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: this wallet is already linked to another Matchlock account
                code: WALLET_OWNED_BY_OTHER
  /auth/wallets/{pubkey}/primary:
    post:
      tags: [auth]
      operationId: postWalletPrimary
      summary: Set primary wallet
      description: Designates the specified linked wallet as the user's primary wallet.
      parameters:
        - $ref: "#/components/parameters/WagerPubkey"
      responses:
        "204":
          description: Primary wallet updated (no content)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Wallet not linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: wallet not linked
                code: WALLET_NOT_FOUND
        "500":
          $ref: "#/components/responses/AuthError"
  /auth/wallets/{pubkey}:
    delete:
      tags: [auth]
      operationId: deleteWallet
      summary: Unlink a wallet
      description: Removes the wallet link from the user's account.
      parameters:
        - $ref: "#/components/parameters/WagerPubkey"
      responses:
        "204":
          description: Wallet unlinked (no content)
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          description: Wallet not linked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: wallet not linked
                code: WALLET_NOT_FOUND
        "500":
          $ref: "#/components/responses/AuthError"
  /users/lookup:
    get:
      tags: [auth]
      operationId: getUserLookup
      summary: Lookup user by email
      description: |
        Resolves an email address to a user profile for challenge invitations.
        Returns whether the user has an account and their primary wallet address.
      parameters:
        - $ref: "#/components/parameters/EmailQuery"
      responses:
        "200":
          description: User lookup result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserLookup"
        "400":
          description: Invalid email
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: invalid email address
                code: INVALID_EMAIL
        "401":
          $ref: "#/components/responses/Unauthorized"
  /leaderboard:
    get:
      tags: [leaderboard]
      operationId: getLeaderboard
      summary: List ranked leaderboard entries
      description: |
        Returns leaderboard rows ordered by net PnL descending, then wins, total volume,
        and user id for deterministic ranking.
      parameters:
        - $ref: "#/components/parameters/OffsetQuery"
        - $ref: "#/components/parameters/LeaderboardLimitQuery"
      responses:
        "200":
          description: Leaderboard page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardPage"
        "500":
          $ref: "#/components/responses/InternalError"
  /leaderboard/me:
    get:
      tags: [leaderboard]
      operationId: getMyLeaderboardRank
      summary: Get the authenticated user's leaderboard row and rank
      responses:
        "200":
          description: Leaderboard entry or empty rank if the user has no settled wagers yet
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/LeaderboardEntry"
                  - type: object
                    required: [rank, total_wagers]
                    properties:
                      rank:
                        type: "null"
                      total_wagers:
                        type: integer
                        example: 0
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /leaderboard/stats:
    get:
      tags: [leaderboard]
      operationId: getLeaderboardStats
      summary: Get leaderboard aggregates
      description: Returns aggregates across all leaderboard rows, not just the first page.
      responses:
        "200":
          description: Leaderboard aggregates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardStats"
        "500":
          $ref: "#/components/responses/InternalError"
  /leaderboard/wagers/{pubkey}/sync:
    post:
      tags: [leaderboard]
      operationId: syncLeaderboardSettlement
      summary: Sync a settled wager into the leaderboard
      description: |
        Reconciles a settled on-chain wager into leaderboard standings. Safe to call multiple
        times; each wager is counted at most once.
      parameters:
        - $ref: "#/components/parameters/WagerPubkey"
        - $ref: "#/components/parameters/TxSignatureQuery"
      responses:
        "200":
          description: Wager has been processed for leaderboard purposes
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeaderboardSyncResponse"
        "400":
          $ref: "#/components/responses/InvalidPubkey"
        "404":
          $ref: "#/components/responses/WagerNotFound"
        "409":
          description: Wager is not settled or final score is unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "500":
          $ref: "#/components/responses/InternalError"
  /challenges/invites:
    post:
      tags: [challenges]
      operationId: postChallengeInvite
      summary: Create a challenge invite
      description: |
        Creates a peer-to-peer challenge invitation. The recipient receives an email
        notification. The invite can optionally include an on-chain wager pubkey.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInviteRequest"
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteView"
        "400":
          description: Invalid invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: recipient cannot be yourself
                code: INVALID_INVITE
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InviteError"
    get:
      tags: [challenges]
      operationId: listChallengeInvites
      summary: List challenge invites
      description: |
        Returns all invites sent to or created by the authenticated user.
        Enriches each invite with on-chain wager status (accepted if matched/settled).
      responses:
        "200":
          description: Invite list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InviteView"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InviteError"
...
  /challenges/invites/{id}/wager:
    post:
      tags: [challenges]
      operationId: postChallengeInviteWager
      summary: Attach wager to invite
      description: |
        Links an on-chain wager to a challenge invite. Only the maker (creator)
        can attach the wager. The wager's maker wallet must belong to the authenticated user.
      parameters:
        - $ref: "#/components/parameters/InviteId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AttachWagerRequest"
      responses:
        "200":
          description: Wager attached
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InviteView"
        "400":
          description: Invalid wager or invite
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                invalidWager:
                  summary: Invalid wager pubkey
                  value:
                    error: invalid wager pubkey
                    code: INVALID_WAGER
                wagerNotFound:
                  summary: Wager not on chain
                  value:
                    error: wager not found on chain
                    code: WAGER_NOT_FOUND
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InviteError"
components:
  parameters:
    MatchId:
      name: id
      in: path
      required: true
      description: Match identifier (decimal fixture id string)
      schema:
        type: string
        example: "17952170"
    MatchIdQuery:
      name: match_id
      in: query
      required: false
      description: Filter by match fixture id
      schema:
        type: string
        example: "17952170"
...
    StatusQuery:
      name: status
      in: query
      required: false
      description: Filter by wager status
      schema:
        $ref: "#/components/schemas/WagerStatus"
...
  schemas:
...
    Error:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
          description: Human-readable message
        code:
          type: string
          description: Machine-readable error code
  responses:
    CacheError:
      description: Redis cache error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: failed to list matches
            code: CACHE_ERROR
...
    InvalidMatchId:
      description: Missing or empty match id
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: match id is required
            code: INVALID_MATCH_ID
    MatchNotFound:
      description: Match not in cache
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: match not found
            code: MATCH_NOT_FOUND
...